feat: ✨ Monitor events in finality (#161)
