+++ api-augment/src/interfaces/augment-api-runtime.ts
@@ -48,6 +48,7 @@ import type {
   QueryBspConfirmChunksToProveForFileError,
   QueryFileEarliestVolunteerBlockError,
   RandomnessOutput,
+  StorageProviderId,
   TrieRemoveMutation
 } from "@storagehub/api-augment/interfaces/storagehubclient";
 
@@ -410,6 +411,13 @@ declare module "@polkadot/api-base/types/calls" {
         (
           bspId: BackupStorageProviderId | string | Uint8Array
         ) => Observable<Result<BackupStorageProvider, GetBspInfoError>>
+      >;
+      /**
+       * Get the Storage Provider ID for a given Account ID.
+       **/
+      getStorageProviderId: AugmentedCall<
+        ApiType,
+        (who: AccountId | string | Uint8Array) => Observable<Option<StorageProviderId>>
       >;
       /**
        * Generic call
