+++ client/blockchain-service/src/events.rs
@@ -1,7 +1,7 @@
 use sc_network::Multiaddr;
 use shc_actors_framework::event_bus::{EventBus, EventBusMessage, ProvidesEventBus};
 use shc_common::types::{
-    BlockNumber, BucketId, FileKey, FileLocation, Fingerprint, PeerIds, ProviderId,
+    BlockNumber, BucketId, FileKey, FileLocation, Fingerprint, ForestRoot, PeerIds, ProviderId,
     RandomnessOutput, StorageData, TrieRemoveMutation,
 };
 use sp_core::H256;
@@ -62,12 +62,6 @@ pub struct AcceptedBspVolunteer {
 
 impl EventBusMessage for AcceptedBspVolunteer {}
 
-// TODO: use proper types
-#[derive(Debug, Clone)]
-pub struct StorageRequestRevoked {
-    pub location: String,
-}
-
 #[derive(Debug, Clone)]
 pub struct ProcessSubmitProofRequest {
     pub provider_id: ProviderId,
@@ -84,8 +78,6 @@ pub struct ProcessConfirmStoringRequest {
     pub forest_root_write_tx: Arc<Mutex<Option<oneshot::Sender<()>>>>,
 }
 
-impl EventBusMessage for StorageRequestRevoked {}
-
 /// Slashable Provider event.
 ///
 /// This event is emitted when a provider is marked as slashable by the runtime.
@@ -97,15 +89,32 @@ pub struct SlashableProvider {
 
 impl EventBusMessage for SlashableProvider {}
 
+/// Mutations applied event in a finalised block.
+///
+/// This event is emitted when a finalised block is received by the Blockchain service,
+/// in which there is a `MutationsApplied` event for one of the providers that this node is tracking.
+#[derive(Debug, Clone)]
+pub struct FinalisedMutationsApplied {
+    pub provider_id: ProviderId,
+    pub mutations: Vec<(ForestRoot, TrieRemoveMutation)>,
+    pub new_root: H256,
+}
+
+impl EventBusMessage for FinalisedMutationsApplied {}
+
+/// The event bus provider for the BlockchainService actor.
+///
+/// It holds the event buses for the different events that the BlockchainService actor
+/// can emit.
 #[derive(Clone, Default)]
 pub struct BlockchainServiceEventBusProvider {
     new_challenge_seed_event_bus: EventBus<NewChallengeSeed>,
     new_storage_request_event_bus: EventBus<NewStorageRequest>,
     accepted_bsp_volunteer_event_bus: EventBus<AcceptedBspVolunteer>,
-    storage_request_revoked_event_bus: EventBus<StorageRequestRevoked>,
     process_submit_proof_request_event_bus: EventBus<ProcessSubmitProofRequest>,
     process_confirm_storage_request_event_bus: EventBus<ProcessConfirmStoringRequest>,
     slashable_provider_event_bus: EventBus<SlashableProvider>,
+    finalised_mutations_applied_event_bus: EventBus<FinalisedMutationsApplied>,
 }
 
 impl EventBusMessage for ProcessSubmitProofRequest {}
@@ -116,10 +125,10 @@ impl BlockchainServiceEventBusProvider {
             new_challenge_seed_event_bus: EventBus::new(),
             new_storage_request_event_bus: EventBus::new(),
             accepted_bsp_volunteer_event_bus: EventBus::new(),
-            storage_request_revoked_event_bus: EventBus::new(),
             process_submit_proof_request_event_bus: EventBus::new(),
             process_confirm_storage_request_event_bus: EventBus::new(),
             slashable_provider_event_bus: EventBus::new(),
+            finalised_mutations_applied_event_bus: EventBus::new(),
         }
     }
 }
@@ -144,12 +153,6 @@ impl ProvidesEventBus<AcceptedBspVolunteer> for BlockchainServiceEventBusProvide
     }
 }
 
-impl ProvidesEventBus<StorageRequestRevoked> for BlockchainServiceEventBusProvider {
-    fn event_bus(&self) -> &EventBus<StorageRequestRevoked> {
-        &self.storage_request_revoked_event_bus
-    }
-}
-
 impl ProvidesEventBus<ProcessSubmitProofRequest> for BlockchainServiceEventBusProvider {
     fn event_bus(&self) -> &EventBus<ProcessSubmitProofRequest> {
         &self.process_submit_proof_request_event_bus
@@ -165,5 +168,11 @@ impl ProvidesEventBus<ProcessConfirmStoringRequest> for BlockchainServiceEventBu
 impl ProvidesEventBus<SlashableProvider> for BlockchainServiceEventBusProvider {
     fn event_bus(&self) -> &EventBus<SlashableProvider> {
         &self.slashable_provider_event_bus
+    }
+}
+
+impl ProvidesEventBus<FinalisedMutationsApplied> for BlockchainServiceEventBusProvider {
+    fn event_bus(&self) -> &EventBus<FinalisedMutationsApplied> {
+        &self.finalised_mutations_applied_event_bus
     }
 }
