+++ client/blockchain-service/src/handler.rs
@@ -1,95 +1,45 @@
-// This file is part of Substrate.
-
-// Copyright (C) Parity Technologies (UK) Ltd.
-// SPDX-License-Identifier: GPL-3.0-or-later WITH Classpath-exception-2.0
-
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU General Public License for more details.
-
-// You should have received a copy of the GNU General Public License
-// along with this program. If not, see <https://www.gnu.org/licenses/>.
-
 use std::{
-    collections::{BTreeMap, VecDeque},
+    collections::{BTreeMap, BTreeSet, VecDeque},
     str::FromStr,
     sync::Arc,
 };
 
-use anyhow::Result;
-use codec::{Decode, Encode};
-use cumulus_primitives_core::BlockT;
-use frame_support::{StorageHasher, Twox128};
-use futures::{prelude::*, stream::select};
-use lazy_static::lazy_static;
+use futures::prelude::*;
 use log::{debug, trace, warn};
 use pallet_storage_providers_runtime_api::{GetBspInfoError, StorageProvidersApi};
-use polkadot_runtime_common::BlockHashCount;
 use sc_client_api::{
-    BlockBackend, BlockImportNotification, BlockchainEvents, HeaderBackend, StorageKey,
-    StorageProvider,
+    BlockImportNotification, BlockchainEvents, FinalityNotification, HeaderBackend,
 };
 use sc_network::Multiaddr;
 use sc_service::RpcHandlers;
 use sc_tracing::tracing::{error, info};
-use serde_json::Number;
 use shc_actors_framework::actor::{Actor, ActorEventLoop};
 use shc_common::types::{Fingerprint, RandomnessOutput, TrieRemoveMutation, BCSV_KEY_TYPE};
 use shp_file_metadata::FileKey;
 use sp_api::ProvideRuntimeApi;
-use sp_core::{Blake2Hasher, Hasher, H256};
+use sp_core::H256;
 use sp_keystore::{Keystore, KeystorePtr};
-use sp_runtime::{
-    generic::{self, SignedPayload},
-    traits::Header,
-    AccountId32, SaturatedConversion,
-};
-use storage_hub_runtime::{RuntimeEvent, SignedExtra, UncheckedExtrinsic};
-use substrate_frame_rpc_system::AccountNonceApi;
-use tokio::sync::{oneshot::error::TryRecvError, Mutex};
+use sp_runtime::{traits::Header, AccountId32, SaturatedConversion};
+use storage_hub_runtime::RuntimeEvent;
 
 use pallet_file_system_runtime_api::{
     FileSystemApi, QueryBspConfirmChunksToProveForFileError, QueryFileEarliestVolunteerBlockError,
 };
 use pallet_proofs_dealer_runtime_api::{
-    GetChallengePeriodError, GetCheckpointChallengesError, GetLastTickProviderSubmittedProofError,
-    ProofsDealerApi,
+    GetCheckpointChallengesError, GetLastTickProviderSubmittedProofError, ProofsDealerApi,
 };
 use shc_common::types::{BlockNumber, ParachainClient, ProviderId};
 
 use crate::{
     commands::BlockchainServiceCommand,
     events::{
-        AcceptedBspVolunteer, BlockchainServiceEventBusProvider, NewChallengeSeed,
-        NewStorageRequest, ProcessConfirmStoringRequest, ProcessSubmitProofRequest,
-        SlashableProvider,
+        AcceptedBspVolunteer, BlockchainServiceEventBusProvider, FinalisedMutationsApplied,
+        NewChallengeSeed, NewStorageRequest, SlashableProvider,
     },
     transaction::SubmittedTransaction,
-    types::{EventsVec, Extrinsic},
 };
 
-const LOG_TARGET: &str = "blockchain-service";
-
-lazy_static! {
-    // Would be cool to be able to do this...
-    // let events_storage_key = frame_system::Events::<storage_hub_runtime::Runtime>::hashed_key();
-
-    // Static and lazily initialised `events_storage_key`
-    static ref EVENTS_STORAGE_KEY: Vec<u8> = {
-        let key = [
-            Twox128::hash(b"System").to_vec(),
-            Twox128::hash(b"Events").to_vec(),
-        ]
-        .concat();
-        key
-    };
-}
+pub(crate) const LOG_TARGET: &str = "blockchain-service";
 
 pub struct SubmitProofRequest {
     pub provider_id: ProviderId,
@@ -132,20 +82,21 @@ impl ConfirmStoringRequest {
 /// extrinsics, and the [`Keystore`] to sign the extrinsics.
 pub struct BlockchainService {
     /// The event bus provider.
-    event_bus_provider: BlockchainServiceEventBusProvider,
+    pub(crate) event_bus_provider: BlockchainServiceEventBusProvider,
     /// The parachain client. Used to interact with the runtime.
-    client: Arc<ParachainClient>,
+    pub(crate) client: Arc<ParachainClient>,
     /// The keystore. Used to sign extrinsics.
-    keystore: KeystorePtr,
+    pub(crate) keystore: KeystorePtr,
     /// The RPC handlers. Used to send extrinsics.
-    rpc_handlers: Arc<RpcHandlers>,
+    pub(crate) rpc_handlers: Arc<RpcHandlers>,
     /// Nonce counter for the extrinsics.
-    nonce_counter: u32,
+    pub(crate) nonce_counter: u32,
     /// A registry of waiters for a block number.
-    wait_for_block_request_by_number: BTreeMap<BlockNumber, Vec<tokio::sync::oneshot::Sender<()>>>,
+    pub(crate) wait_for_block_request_by_number:
+        BTreeMap<BlockNumber, Vec<tokio::sync::oneshot::Sender<()>>>,
     /// A list of Provider IDs that this node has to pay attention to submit proofs for.
     /// This could be a BSP or a list of buckets that an MSP has.
-    provider_ids: Vec<ProviderId>,
+    pub(crate) provider_ids: BTreeSet<ProviderId>,
     /// A lock to prevent multiple tasks from writing to the runtime forest root (send transactions) at the same time.
     /// This is an [`Arc<Mutex<Option<T>>>`] (in this case [`oneshot::Sender<()>`]) instead of just
     /// T so that we can keep using the current actors event bus (emit) which requires Clone on the
@@ -154,11 +105,76 @@ pub struct BlockchainService {
     /// Also, this is a oneshot channel instead of a regular mutex because we want to "lock" in 1
     /// thread (blockchain service) and unlock it at the end of the spawned task. The alternative
     /// would be to send a [`MutexGuard`].
-    forest_root_write_lock: Option<tokio::sync::oneshot::Receiver<()>>,
+    pub(crate) forest_root_write_lock: Option<tokio::sync::oneshot::Receiver<()>>,
     /// A list of [SubmitProofRequest] that are waiting to be processed (sent to the runtime).
-    pending_submit_proof: VecDeque<SubmitProofRequest>,
+    pub(crate) pending_submit_proof: VecDeque<SubmitProofRequest>,
     /// A list of [ConfirmStoringRequest] that are waiting to be processed (sent to the runtime).
-    pending_confirm_storing: VecDeque<ConfirmStoringRequest>,
+    pub(crate) pending_confirm_storing: VecDeque<ConfirmStoringRequest>,
+}
+
+/// Event loop for the BlockchainService actor.
+pub struct BlockchainServiceEventLoop {
+    receiver: sc_utils::mpsc::TracingUnboundedReceiver<BlockchainServiceCommand>,
+    actor: BlockchainService,
+}
+
+/// Merged event loop message for the BlockchainService actor.
+enum MergedEventLoopMessage<Block>
+where
+    Block: cumulus_primitives_core::BlockT,
+{
+    Command(BlockchainServiceCommand),
+    BlockImportNotification(BlockImportNotification<Block>),
+    FinalityNotification(FinalityNotification<Block>),
+}
+
+/// Implement the ActorEventLoop trait for the BlockchainServiceEventLoop.
+impl ActorEventLoop<BlockchainService> for BlockchainServiceEventLoop {
+    fn new(
+        actor: BlockchainService,
+        receiver: sc_utils::mpsc::TracingUnboundedReceiver<BlockchainServiceCommand>,
+    ) -> Self {
+        Self { actor, receiver }
+    }
+
+    async fn run(mut self) {
+        info!(target: LOG_TARGET, "BlockchainService starting up!");
+
+        // Import notification stream to be notified of new blocks.
+        // This will notify us when sync to the latest block, or if there is a re-org.
+        let block_import_notification_stream = self.actor.client.import_notification_stream();
+
+        // Finality notification stream to be notified of blocks being finalised.
+        let finality_notification_stream = self.actor.client.finality_notification_stream();
+
+        // Merging notification streams with command stream.
+        let mut merged_stream = stream::select_all(vec![
+            self.receiver.map(MergedEventLoopMessage::Command).boxed(),
+            block_import_notification_stream
+                .map(MergedEventLoopMessage::BlockImportNotification)
+                .boxed(),
+            finality_notification_stream
+                .map(MergedEventLoopMessage::FinalityNotification)
+                .boxed(),
+        ]);
+
+        // Process incoming messages.
+        while let Some(notification) = merged_stream.next().await {
+            match notification {
+                MergedEventLoopMessage::Command(command) => {
+                    self.actor.handle_message(command).await;
+                }
+                MergedEventLoopMessage::BlockImportNotification(notification) => {
+                    self.actor
+                        .handle_block_import_notification(notification)
+                        .await;
+                }
+                MergedEventLoopMessage::FinalityNotification(notification) => {
+                    self.actor.handle_finality_notification(notification).await;
+                }
+            };
+        }
+    }
 }
 
 /// Implement the Actor trait for the BlockchainService actor.
@@ -518,76 +534,6 @@ impl Actor for BlockchainService {
     }
 }
 
-/// Event loop for the BlockchainService actor.
-pub struct BlockchainServiceEventLoop {
-    receiver: sc_utils::mpsc::TracingUnboundedReceiver<BlockchainServiceCommand>,
-    actor: BlockchainService,
-}
-
-/// Merged event loop message for the BlockchainService actor.
-enum MergedEventLoopMessage<Block>
-where
-    Block: cumulus_primitives_core::BlockT,
-{
-    Command(BlockchainServiceCommand),
-    BlockNotification(BlockImportNotification<Block>),
-}
-
-/// Implement the ActorEventLoop trait for the BlockchainServiceEventLoop.
-impl ActorEventLoop<BlockchainService> for BlockchainServiceEventLoop {
-    fn new(
-        actor: BlockchainService,
-        receiver: sc_utils::mpsc::TracingUnboundedReceiver<BlockchainServiceCommand>,
-    ) -> Self {
-        Self { actor, receiver }
-    }
-
-    async fn run(mut self) {
-        info!(target: LOG_TARGET, "BlockchainService starting up!");
-
-        // Import notification stream to be notified of new blocks.
-        let notification_stream = self.actor.client.import_notification_stream();
-
-        // Merging notification stream with command stream.
-        let mut merged_stream = select(
-            self.receiver.map(MergedEventLoopMessage::Command),
-            notification_stream.map(MergedEventLoopMessage::BlockNotification),
-        );
-
-        // Process incoming messages.
-        while let Some(notification) = merged_stream.next().await {
-            match notification {
-                MergedEventLoopMessage::Command(command) => {
-                    self.actor.handle_message(command).await;
-                }
-                MergedEventLoopMessage::BlockNotification(notification) => {
-                    self.actor.handle_block_notification(notification).await;
-                }
-            };
-        }
-    }
-}
-
-/// The output of an RPC transaction.
-pub struct RpcExtrinsicOutput {
-    /// Hash of the extrinsic.
-    pub hash: H256,
-    /// The output string of the transaction if any.
-    pub result: String,
-    /// An async receiver if data will be returned via a callback.
-    pub receiver: tokio::sync::mpsc::Receiver<String>,
-}
-
-impl std::fmt::Debug for RpcExtrinsicOutput {
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        write!(
-            f,
-            "RpcExtrinsicOutput {{ hash: {:?}, result: {:?}, receiver }}",
-            self.hash, self.result
-        )
-    }
-}
-
 impl BlockchainService {
     /// Create a new [`BlockchainService`].
     pub fn new(
@@ -602,7 +548,7 @@ impl BlockchainService {
             event_bus_provider: BlockchainServiceEventBusProvider::new(),
             nonce_counter: 0,
             wait_for_block_request_by_number: BTreeMap::new(),
-            provider_ids: Vec::new(),
+            provider_ids: BTreeSet::new(),
             forest_root_write_lock: None,
             pending_submit_proof: VecDeque::new(),
             pending_confirm_storing: VecDeque::new(),
@@ -610,7 +556,7 @@ impl BlockchainService {
     }
 
     /// Handle a block import notification.
-    async fn handle_block_notification<Block>(
+    async fn handle_block_import_notification<Block>(
         &mut self,
         notification: BlockImportNotification<Block>,
     ) where
@@ -622,38 +568,14 @@ impl BlockchainService {
         debug!(target: LOG_TARGET, "Import notification #{}: {}", block_number, block_hash);
 
         // Notify all tasks waiting for this block number (or lower).
-        let mut keys_to_remove = Vec::new();
-
-        for (block_number, waiters) in self
-            .wait_for_block_request_by_number
-            .range_mut(..=block_number)
-        {
-            keys_to_remove.push(*block_number);
-            for waiter in waiters.drain(..) {
-                match waiter.send(()) {
-                    Ok(_) => {}
-                    Err(_) => {
-                        error!(target: LOG_TARGET, "Failed to notify task about block number.");
-                    }
-                }
-            }
-        }
-
-        for key in keys_to_remove {
-            self.wait_for_block_request_by_number.remove(&key);
-        }
+        self.notify_import_block_number(block_number);
 
         // We query the [`BlockchainService`] account nonce at this height
         // and update our internal counter if it's smaller than the result.
-        let pub_key = Self::caller_pub_key(self.keystore.clone());
-        let latest_nonce = self
-            .client
-            .runtime_api()
-            .account_nonce(block_hash, pub_key.into())
-            .expect("Fetching account nonce works; qed");
-        if latest_nonce > self.nonce_counter {
-            self.nonce_counter = latest_nonce
-        }
+        self.check_nonce(block_hash);
+
+        // Get provider IDs linked to keys in this node's keystore.
+        self.get_provider_ids(block_hash);
 
         // Process pending requests that update the forest root.
         self.check_pending_forest_root_writes();
@@ -701,7 +623,7 @@ impl BlockchainService {
                                 if self.keystore.has_keys(&[(account.clone(), BCSV_KEY_TYPE)]) {
                                     // If so, add the Provider ID to the list of Providers that this node is monitoring.
                                     info!(target: LOG_TARGET, "New Provider ID to monitor [{:?}] for account [{:?}]", provider_id, account);
-                                    self.provider_ids.push(provider_id);
+                                    self.provider_ids.insert(provider_id);
                                 }
                             }
                         }
@@ -802,357 +724,53 @@ impl BlockchainService {
         }
     }
 
-    /// Send an extrinsic to this node using an RPC call.
-    async fn send_extrinsic(
+    /// Handle a finality notification.
+    async fn handle_finality_notification<Block>(
         &mut self,
-        call: impl Into<storage_hub_runtime::RuntimeCall>,
-    ) -> Result<RpcExtrinsicOutput> {
-        debug!(target: LOG_TARGET, "Sending extrinsic to the runtime");
+        notification: FinalityNotification<Block>,
+    ) where
+        Block: cumulus_primitives_core::BlockT<Hash = H256>,
+    {
+        let block_hash: H256 = notification.hash;
+        let block_number: BlockNumber = (*notification.header.number()).saturated_into();
 
-        // Get the nonce for the caller and increment it for the next transaction.
-        // TODO: Handle nonce overflow.
-        let nonce = self.nonce_counter;
+        debug!(target: LOG_TARGET, "Finality notification #{}: {}", block_number, block_hash);
 
-        // Construct the extrinsic.
-        let extrinsic = self.construct_extrinsic(self.client.clone(), call, nonce);
-
-        // Generate a unique ID for this query.
-        let id_hash = Blake2Hasher::hash(&extrinsic.encode());
-        // TODO: Consider storing the ID in a hashmap if later retrieval is needed.
-
-        let (result, rx) = self
-            .rpc_handlers
-            .rpc_query(&format!(
-                r#"{{
-                    "jsonrpc": "2.0",
-                    "method": "author_submitAndWatchExtrinsic",
-                    "params": ["0x{}"],
-                    "id": {:?}
-                }}"#,
-                array_bytes::bytes2hex("", &extrinsic.encode()),
-                array_bytes::bytes2hex("", &id_hash.as_bytes())
-            ))
-            .await
-            .expect("Sending query failed even when it is correctly formatted as JSON-RPC; qed");
-
-        let json: serde_json::Value =
-            serde_json::from_str(&result).expect("the result can only be a JSONRPC string; qed");
-        let error = json
-            .as_object()
-            .expect("JSON result is always an object; qed")
-            .get("error");
-
-        if let Some(error) = error {
-            // TODO: Consider how to handle a low nonce error, and retry.
-            return Err(anyhow::anyhow!("Error in RPC call: {}", error.to_string()));
-        }
-
-        // Only update nonce after we are sure no errors
-        // occurred submitting the extrinsic.
-        self.nonce_counter += 1;
-
-        Ok(RpcExtrinsicOutput {
-            hash: id_hash,
-            result,
-            receiver: rx,
-        })
-    }
-
-    /// Construct an extrinsic that can be applied to the runtime.
-    pub fn construct_extrinsic(
-        &self,
-        client: Arc<ParachainClient>,
-        function: impl Into<storage_hub_runtime::RuntimeCall>,
-        nonce: u32,
-    ) -> UncheckedExtrinsic {
-        let function = function.into();
-        let current_block_hash = client.info().best_hash;
-        let current_block = client.info().best_number.saturated_into();
-        let genesis_block = client
-            .hash(0)
-            .expect("Failed to get genesis block hash, always present; qed")
-            .expect("Genesis block hash should never not be on-chain; qed");
-        let period = BlockHashCount::get()
-            .checked_next_power_of_two()
-            .map(|c| c / 2)
-            .unwrap_or(2) as u64;
-        // TODO: Consider tipping the transaction.
-        let tip = 0;
-        let extra: SignedExtra = (
-        frame_system::CheckNonZeroSender::<storage_hub_runtime::Runtime>::new(),
-        frame_system::CheckSpecVersion::<storage_hub_runtime::Runtime>::new(),
-        frame_system::CheckTxVersion::<storage_hub_runtime::Runtime>::new(),
-        frame_system::CheckGenesis::<storage_hub_runtime::Runtime>::new(),
-        frame_system::CheckEra::<storage_hub_runtime::Runtime>::from(generic::Era::mortal(
-            period,
-            current_block,
-        )),
-        frame_system::CheckNonce::<storage_hub_runtime::Runtime>::from(nonce),
-        frame_system::CheckWeight::<storage_hub_runtime::Runtime>::new(),
-        pallet_transaction_payment::ChargeTransactionPayment::<storage_hub_runtime::Runtime>::from(
-            tip,
-        ),
-        cumulus_primitives_storage_weight_reclaim::StorageWeightReclaim::<
-            storage_hub_runtime::Runtime,
-        >::new(),
-    );
-
-        let raw_payload = SignedPayload::from_raw(
-            function.clone(),
-            extra.clone(),
-            (
-                (),
-                storage_hub_runtime::VERSION.spec_version,
-                storage_hub_runtime::VERSION.transaction_version,
-                genesis_block,
-                current_block_hash,
-                (),
-                (),
-                (),
-                (),
-            ),
-        );
-
-        let caller_pub_key = Self::caller_pub_key(self.keystore.clone());
-
-        // Sign the payload.
-        let signature = raw_payload
-            .using_encoded(|e| self.keystore.sr25519_sign(BCSV_KEY_TYPE, &caller_pub_key, e))
-            .expect("The payload is always valid and should be possible to sign; qed")
-            .expect("They key type and public key are valid because we just extracted them from the keystore; qed");
-
-        // Construct the extrinsic.
-        UncheckedExtrinsic::new_signed(
-            function.clone(),
-            storage_hub_runtime::Address::Id(<sp_core::sr25519::Public as Into<
-                storage_hub_runtime::AccountId,
-            >>::into(caller_pub_key)),
-            polkadot_primitives::Signature::Sr25519(signature.clone()),
-            extra.clone(),
-        )
-    }
-
-    // Getting signer public key.
-    pub fn caller_pub_key(keystore: KeystorePtr) -> sp_core::sr25519::Public {
-        let caller_pub_key = keystore.sr25519_public_keys(BCSV_KEY_TYPE).pop().expect(
-            format!(
-                "There should be at least one sr25519 key in the keystore with key type '{:?}' ; qed",
-                BCSV_KEY_TYPE
-            )
-            .as_str(),
-        );
-        caller_pub_key
-    }
-
-    /// Get an extrinsic from a block.
-    async fn get_extrinsic_from_block(
-        &self,
-        block_hash: H256,
-        extrinsic_hash: H256,
-    ) -> Result<Extrinsic> {
-        // Get the block.
-        let block = self
-            .client
-            .block(block_hash)
-            .expect("Failed to get block. This shouldn't be possible for known existing block hash; qed")
-            .expect("Block returned None for known existing block hash. This shouldn't be the case for a block known to have at least one transaction; qed");
-
-        // Get the extrinsics.
-        let extrinsics = block.block.extrinsics();
-
-        // Find the extrinsic index in the block.
-        let extrinsic_index = extrinsics
-            .iter()
-            .position(|e| {
-                let hash = Blake2Hasher::hash(&e.encode());
-                hash == extrinsic_hash
-            })
-            .expect("Extrinsic not found in block. This shouldn't be possible if we're looking into a block for which we got confirmation that the extrinsic was included; qed");
-
-        // Get the events from storage.
-        let events_in_block = self.get_events_storage_element(block_hash)?;
-
-        // Filter the events for the extrinsic.
-        // Each event record is composed of the `phase`, `event` and `topics` fields.
-        // We are interested in those events whose `phase` is equal to `ApplyExtrinsic` with the index of the extrinsic.
-        // For more information see: https://polkadot.js.org/docs/api/cookbook/blocks/#how-do-i-map-extrinsics-to-their-events
-        let events = events_in_block
-            .into_iter()
-            .filter(|ev| ev.phase == frame_system::Phase::ApplyExtrinsic(extrinsic_index as u32))
-            .collect();
-
-        // Construct the extrinsic.
-        Ok(Extrinsic {
-            hash: extrinsic_hash,
-            block_hash,
-            events,
-        })
-    }
-
-    /// Unwatch an extrinsic.
-    async fn unwatch_extrinsic(&self, subscription_id: Number) -> Result<String> {
-        let (result, _rx) = self
-            .rpc_handlers
-            .rpc_query(&format!(
-                r#"{{
-                    "jsonrpc": "2.0",
-                    "method": "author_unwatchExtrinsic",
-                    "params": [{}],
-                    "id": {}
-                }}"#,
-                subscription_id, subscription_id
-            ))
-            .await
-            .expect("Sending query failed even when it is correctly formatted as JSON-RPC; qed");
-
-        let json: serde_json::Value =
-            serde_json::from_str(&result).expect("the result can only be a JSONRPC string; qed");
-        let unwatch_result = json
-            .as_object()
-            .expect("JSON result is always an object; qed")
-            .get("result");
-
-        if let Some(unwatch_result) = unwatch_result {
-            if unwatch_result
-                .as_bool()
-                .expect("Result is always a boolean; qed")
-            {
-                debug!(target: LOG_TARGET, "Extrinsic unwatched successfully");
-            } else {
-                return Err(anyhow::anyhow!("Failed to unwatch extrinsic"));
+        // Get events from storage.
+        match self.get_events_storage_element(block_hash) {
+            Ok(block_events) => {
+                // Process the events.
+                for ev in block_events {
+                    match ev.event.clone() {
+                        // New storage request event coming from pallet-file-system.
+                        RuntimeEvent::ProofsDealer(
+                            pallet_proofs_dealer::Event::MutationsApplied {
+                                provider,
+                                mutations,
+                                new_root,
+                            },
+                        ) => {
+                            // Check if the provider ID is one of the provider IDs this node is tracking.
+                            if self.provider_ids.contains(&provider) {
+                                self.emit(FinalisedMutationsApplied {
+                                    provider_id: provider.clone(),
+                                    mutations: mutations.clone(),
+                                    new_root: new_root.clone(),
+                                })
+                            }
+                        }
+                        // Ignore all other events.
+                        _ => {}
+                    }
+                }
             }
-        } else {
-            return Err(anyhow::anyhow!("Failed to unwatch extrinsic"));
-        }
-
-        Ok(result)
-    }
-
-    /// Get the events storage element in a block.
-    fn get_events_storage_element(&self, block_hash: H256) -> Result<EventsVec> {
-        // Get the events storage.
-        let raw_storage_opt = self
-            .client
-            .storage(block_hash, &StorageKey(EVENTS_STORAGE_KEY.clone()))
-            .expect("Failed to get Events storage element");
-
-        // Decode the events storage.
-        if let Some(raw_storage) = raw_storage_opt {
-            let block_events = EventsVec::decode(&mut raw_storage.0.as_slice())
-                .expect("Failed to decode Events storage element");
-
-            return Ok(block_events);
-        } else {
-            return Err(anyhow::anyhow!("Failed to get Events storage element"));
-        }
-    }
-
-    /// Check if the challenges tick is one that this provider has to submit a proof for,
-    /// and if so, emit a `NewChallengeSeed` event.
-    fn should_provider_submit_proof(
-        &self,
-        block_hash: &H256,
-        provider_id: &ProviderId,
-        current_tick: &BlockNumber,
-    ) -> bool {
-        let last_tick_provided = match self
-            .client
-            .runtime_api()
-            .get_last_tick_provider_submitted_proof(*block_hash, provider_id)
-        {
-            Ok(last_tick_provided_result) => match last_tick_provided_result {
-                Ok(last_tick_provided) => last_tick_provided,
-                Err(e) => match e {
-                    GetLastTickProviderSubmittedProofError::ProviderNotRegistered => {
-                        debug!(target: LOG_TARGET, "Provider [{:?}] is not registered", provider_id);
-                        return false;
-                    }
-                    GetLastTickProviderSubmittedProofError::ProviderNeverSubmittedProof => {
-                        debug!(target: LOG_TARGET, "Provider [{:?}] does not have an initialised challenge cycle", provider_id);
-                        return false;
-                    }
-                    GetLastTickProviderSubmittedProofError::InternalApiError => {
-                        error!(target: LOG_TARGET, "This should be impossible, we just checked the API error. \nInternal API error while getting last tick Provider [{:?}] submitted a proof for: {:?}", provider_id, e);
-                        return false;
-                    }
-                },
-            },
             Err(e) => {
-                error!(target: LOG_TARGET, "Runtime API error while getting last tick Provider [{:?}] submitted a proof for: {:?}", provider_id, e);
-                return false;
+                // TODO: Handle case where the storage cannot be decoded.
+                // TODO: This would happen if we're parsing a block authored with an older version of the runtime, using
+                // TODO: a node that has a newer version of the runtime, therefore the EventsVec type is different.
+                // TODO: Consider using runtime APIs for getting old data of previous blocks, and this just for current blocks.
+                error!(target: LOG_TARGET, "Failed to get events storage element: {:?}", e);
             }
-        };
-        let provider_challenge_period = match self
-            .client
-            .runtime_api()
-            .get_challenge_period(*block_hash, provider_id)
-        {
-            Ok(provider_challenge_period_result) => match provider_challenge_period_result {
-                Ok(provider_challenge_period) => provider_challenge_period,
-                Err(e) => match e {
-                    GetChallengePeriodError::ProviderNotRegistered => {
-                        debug!(target: LOG_TARGET, "Provider [{:?}] is not registered", provider_id);
-                        return false;
-                    }
-                },
-            },
-            Err(e) => {
-                debug!(target: LOG_TARGET, "Runtime API error while getting challenge period for Provider [{:?}]: {:?}", provider_id, e);
-                return false;
-            }
-        };
-        current_tick == &last_tick_provided.saturating_add(provider_challenge_period)
-    }
-
-    /// Check if there are any pending requests to update the forest root on the runtime, and process them.
-    /// Takes care of prioritizing requests, favoring `SubmitProofRequest` over `ConfirmStoringRequest`.
-    /// This function is called every time a new block is imported and after each request is queued.
-    fn check_pending_forest_root_writes(&mut self) {
-        if let Some(mut rx) = self.forest_root_write_lock.take() {
-            // Note: tasks that get ownership of the lock are responsible for sending a message back when done processing.
-            match rx.try_recv() {
-                // If the channel is empty, means we still need to wait for the current task to finish.
-                Err(TryRecvError::Empty) => {
-                    // If we have a task writing to the runtime, we don't want to start another one.
-                    self.forest_root_write_lock = Some(rx);
-                    trace!(target: LOG_TARGET, "Waiting for current forest root write task to finish");
-                    return;
-                }
-                Ok(_) => {
-                    trace!(target: LOG_TARGET, "Forest root write task finished, lock is released!");
-                }
-                Err(TryRecvError::Closed) => {
-                    error!(target: LOG_TARGET, "Forest root write task channel closed unexpectedly. Lock is released anyway!");
-                }
-            }
-        }
-
-        // At this point we know that the lock is released and we can start processing new requests.
-        let (tx, rx) = tokio::sync::oneshot::channel();
-        self.forest_root_write_lock = Some(rx);
-
-        let forest_root_write_tx = Arc::new(Mutex::new(Some(tx)));
-
-        // If we have a submit proof request, prioritize it.
-        if let Some(request) = self.pending_submit_proof.pop_front() {
-            self.emit(ProcessSubmitProofRequest {
-                seed: request.seed,
-                provider_id: request.provider_id,
-                tick: request.tick,
-                forest_challenges: request.forest_challenges,
-                checkpoint_challenges: request.checkpoint_challenges,
-                forest_root_write_tx,
-            });
-            return;
-        }
-
-        // If we have a confirm storing request, start processing it.
-        if let Some(request) = self.pending_confirm_storing.pop_front() {
-            self.emit(ProcessConfirmStoringRequest {
-                file_key: request.file_key,
-                forest_root_write_tx,
-            });
-            return;
         }
     }
 }
