+++ pallets/providers/src/types.rs
@@ -68,6 +68,15 @@ pub enum StorageProvider<T: Config> {
     MainStorageProvider(MainStorageProvider<T>),
 }
 
+/// Enum that represents a Storage Provider ID. It holds either a BackupStorageProviderId or a MainStorageProviderId,
+/// allowing to operate generically with both types.
+#[derive(Encode, Decode, MaxEncodedLen, TypeInfo, RuntimeDebugNoBound, PartialEq, Eq, Clone)]
+#[scale_info(skip_type_params(T))]
+pub enum StorageProviderId<T: Config> {
+    BackupStorageProvider(BackupStorageProviderId<T>),
+    MainStorageProvider(MainStorageProviderId<T>),
+}
+
 // Type aliases:
 
 /// BalanceOf is the balance type of the runtime.
