+++ types-bundle/src/runtime.ts
@@ -131,6 +131,16 @@ const STORAGE_PROVIDERS_V1: Record<string, DefinitionCall> = {
       }
     ],
     type: "Result<BackupStorageProvider, GetBspInfoError>"
+  },
+  get_storage_provider_id: {
+    description: "Get the Storage Provider ID for a given Account ID.",
+    params: [
+      {
+        name: "who",
+        type: "AccountId"
+      }
+    ],
+    type: "Option<StorageProviderId>"
   }
 };
