+++ types-bundle/src/types.ts
@@ -25,6 +25,7 @@ export const ALL_TYPES: RegistryTypes = {
   RandomnessOutput: "H256",
   TrieRemoveMutation: {},
   BackupStorageProviderId: "H256",
+  MainStorageProviderId: "H256",
   StorageData: "u32",
   MerklePatriciaRoot: "H256",
   ChunkId: "u64",
@@ -36,6 +37,12 @@ export const ALL_TYPES: RegistryTypes = {
     last_capacity_change: "BlockNumber",
     owner_account: "AccountId",
     payment_account: "AccountId"
+  },
+  StorageProviderId: {
+    _enum: {
+      BackupStorageProvider: "BackupStorageProviderId",
+      MainStorageProvider: "MainStorageProviderId"
+    }
   },
   GetLastTickProviderSubmittedProofError: {
     _enum: {
